Detect concurrent modification of cached files
The plugin is now able to detect concurrent modification of the cached files while a script is being executed. This can happen when a script is being started, takes a while to run, and while it is doing so, the developer continues to modify source files. This results in uncertainty about which state of the files the script was executed on. To avoid faulty cache entries, the cache will not be updated in this case.